+++ buy.py
@@ -1,50 +1,49 @@
 from objects import Text_Button, Rect
 from colors import *
 
-def Buy_Button(game, value, name, i):
-	level = value_dict[name].index(value)
-	price = price_dict[name][level]
-	reward = value_dict[name][value + 1]
-	if price == "max":
-		return Null_Button(game, name, i)
-	else:
-		return Buy_Button_Obj(game, price, name, reward, i)
-	return None
 
 
-class Buy_Button_Obj:
-	def __init__(self, game, price, name, reward, i):
-		self.price = price
-		self.game = game
-		self.name = name
-		# self.pic = img_dict[name]
-		buy_pos = (75 + 200 * i, 200 if i < 8 else 500)
-		x, y = buy_pos
-		self.buy = Text_Button(self.game.scr, str(price), x, y, 50, BLACK, WHITE, resp=self.buy)
-		rect_pos = (x / 3, y - 50)
-		x, y = rect_pos
-		self.rect = Rect(self.game.scr, BLUE, x, y, 150, 150)
-		self.blit_buy_button = self.buy.blit
-		self.blit_big_rect = self.rect.blit
-		self.reward = reward
+def Generate_Buttons(game):
+	scr = game.scr
+	prices = game.prices
+	values = game.values
+	levels = game.levels
+	names = game.names
+	out = []
+	for i, name in enumerate(names):
+		current_prices = prices[name]
+		current_level = levels[name]
+		current_values = values[name]
+		price = current_prices[current_level + 1]
+		value = current_values[current_level + 1]
+		out.append(Buy_Button(game, name, price, value, i, current_level + 1))
+	return out
+
+
+class Buy_Button:
+	def __init__(self, *args):
+		self.game, self.name, self.price, self.value, index, self.level = args
+		self.button = Text_Button(self.game.scr, str(self.price), 150, 510, 50, WHITE, DARK_GREEN, resp=self.buy)
+		self.blit = self.button.blit
+		self.__lock = False
+		return
+
+	def lock(self):
+		self.__lock = True
 		return
 
 	def buy(self):
-		self.game.buy(self)
+		if not self.__lock:
+			self.game.buy(self)
 		return
 
-	def blit_img(self):
-		## TODO ##
-		pass
-
-	def blit(self):
-		self.blit_big_rect()
-		self.blit_buy_button()
-		self.blit_img()
+	def update(self, game):
+		self.level += 1
+		if self.level == len(game.values[self.name]):
+			self.lock()
+		else:
+			self.value = game.values[self.name][self.level]
+			self.price = game.prices[self.name][self.level]
+			self.button.update_text(str(self.price))
 		return
 	pass
-
-
-img_arr = []
-value_dict = {"dollars": [1, 5, 10]}
-price_dict = {"dollars": [10, 50, "max"]}